auth done

+++ reviews/views.py
@@ -1,7 +1,8 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
+from rest_framework.permissions import IsAuthenticated
 
 # import models
 from .models import Review
@@ -10,7 +11,10 @@ from .models import Review
 from .serializers.common import ReviewSerializer
 
 class ReviewListView(APIView):
+    permission_classes = (IsAuthenticated, )
+
     def post(self, request):
+        request.data['owner'] = request.user.id
         review_to_add = ReviewSerializer(data=request.data)
         if review_to_add.is_valid():
             review_to_add.save()
@@ -18,6 +22,7 @@ class ReviewListView(APIView):
         return Response(review_to_add.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
 class ReviewDetailView(APIView):
+    permission_classes = (IsAuthenticated, )
 
     def get_review(self, pk):
         try:
@@ -28,5 +33,7 @@ class ReviewDetailView(APIView):
 
     def delete(self, request, pk):
         review_to_delete = self.get_review(pk=pk)
+        if review_to_delete.owner != request.user:
+            raise PermissionDenied(detail="Unauthorised")
         review_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
